naitho-bomb/pyproject.toml
[project]
name = "naitho-bomb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyserial>=3.5",
    "simpleaudio>=1.0.4",
    "pydub>=0.25",
    "audioop-lts>=0.2.2",
]
